added a lot of data seeding into the db

+++ DesktopPurchasingSystem/desktopAppAPI/Models/SupplierModel.cs
@@ -8,6 +8,6 @@ namespace DesktopAppAPI.Models
         public Guid ID { get; set; }
         public string Name { get; set; } = string.Empty;
         public int Delivery_Period_In_Days { get; set; }
-        public int Price_Per_Delivery { get; set; }
+        public float Price_Per_Delivery { get; set; }
     }
 }
